修复bug

+++ src/controller/Assign/index.js
@@ -1,8 +1,7 @@
 import assignModel from '../../model/Assign'
 import sequelize from 'sequelize';
-import bookListModel from '../../model/BookList';
 import JWTDecode from 'jwt-decode';
-import xlsx from 'node-xlsx'
+
 const Op = sequelize.Op;
 
 class AssignController {
@@ -10,31 +9,50 @@ class AssignController {
   }
 
   async getList(ctx) {
-    const {className = '', assignName = '', userName = ''} = ctx.query;
-    ctx.body = await assignModel.getList({
-      className: {[Op.like]: `%${className}%`},
-      assignName: {[Op.like]: `%${assignName}%`},
-      userName: {[Op.like]: `%${userName}%`}
-    })
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {className = '', assignName = '', userName = ''} = ctx.query;
+      ctx.body = await assignModel.getList({
+        className: {[Op.like]: `%${className}%`},
+        assignName: {[Op.like]: `%${assignName}%`},
+        userName: {[Op.like]: `%${userName}%`}
+      })
+    }
   }
 
   async add(ctx) {
-    const {userId, classId, status, assignId} = ctx.request.body;
-    ctx.body = await assignModel.insert({userId, classId, status, assignId});
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {userId, classId, status, assignId} = ctx.request.body;
+      ctx.body = await assignModel.insert({userId, classId, status, assignId});
+    }
   }
 
   async delete(ctx) {
-    const {id} = ctx.params;
-    ctx.body = await assignModel.deleteById({id})
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {id} = ctx.params;
+      ctx.body = await assignModel.deleteById({id})
+    }
+
   }
 
   async update(ctx) {
-    const {id} = ctx.params;
-    const {userId, classId, status, assignId} = ctx.request.body;
-    ctx.body = await assignModel.update({id, userId, classId, status, assignId});
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {id} = ctx.params;
+      const {userId, classId, status, assignId} = ctx.request.body;
+      ctx.body = await assignModel.update({id, userId, classId, status, assignId});
+    }
   }
-
-
 }
 
 export default new AssignController()
