+++ src/controller/Book/index.js
@@ -1,5 +1,7 @@
 import bookModel from '../../model/Book'
 import Sequelize from 'sequelize';
+import JWTDecode from 'jwt-decode';
+import classModel from '../../model/Class';
 
 const Op = Sequelize.Op;
 
@@ -8,24 +10,44 @@ class BookController {
   }
 
   async getList(ctx) {
-    const {bookName = '', ISBN = '', status, toggle} = ctx.query;
-    ctx.body = await bookModel.getList({bookName: {[Op.like]: `%${bookName}%`}, ISBN: {[Op.like]: `%${ISBN}%`}, status, toggle})
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2,3].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {bookName = '', ISBN = '', status, toggle} = ctx.query;
+      ctx.body = await bookModel.getList({bookName: {[Op.like]: `%${bookName}%`}, ISBN: {[Op.like]: `%${ISBN}%`}, status, toggle})
+    }
   }
 
   async add(ctx) {
-    const {bookName, ISBN, status, toggle} = ctx.request.body;
-    ctx.body = await bookModel.insert({bookName, ISBN, status, toggle});
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {bookName, ISBN, status, toggle} = ctx.request.body;
+      ctx.body = await bookModel.insert({bookName, ISBN, status, toggle});
+    }
   }
 
   async delete(ctx) {
-    const {id} = ctx.params;
-    ctx.body = await bookModel.deleteById({id})
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {id} = ctx.params;
+      ctx.body = await bookModel.deleteById({id})
+    }
   }
 
   async update(ctx) {
-    const {id} = ctx.params;
-    const {bookName, ISBN, status, toggle} = ctx.request.body;
-    ctx.body = await bookModel.update({id, bookName, ISBN, status, toggle})
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {id} = ctx.params;
+      const {bookName, ISBN, status, toggle} = ctx.request.body;
+      ctx.body = await bookModel.update({id, bookName, ISBN, status, toggle})
+    }
   }
 
   async getUserBook(ctx) {
@@ -33,8 +55,13 @@ class BookController {
   }
 
   async applyBook(ctx) {
-    const {bookName, ISBN} = ctx.request.body;
-    ctx.body = await bookModel.insert({bookName, ISBN, status: '1', toggle: '0'});
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![3].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {bookName, ISBN} = ctx.request.body;
+      ctx.body = await bookModel.insert({bookName, ISBN, status: '1', toggle: '0'});
+    }
   }
 
   async getBookQuoteList(ctx) {
