+++ src/controller/Quote/index.js
@@ -4,6 +4,7 @@ import JWTDecode from 'jwt-decode';
 import _ from 'lodash';
 import {Book, Quote} from '../../model/model';
 import {seq} from '../../model/seq';
+import classModel from '../../model/Class';
 
 const Op = sequelize.Op;
 
@@ -12,35 +13,65 @@ class QuoteController {
   }
 
   async getList(ctx) {
-    const {bookId, sellerId, price = '', status = ''} = ctx.query;
-    ctx.body = await quoteModel.getList({bookId, sellerId, price: {[Op.like]: `%${price}%`}, status});
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {bookId, sellerId, price = '', status = '',subscriptionId=''} = ctx.query;
+      ctx.body = await quoteModel.getList({bookId, sellerId, price: {[Op.like]: `%${price}%`}, status,subscriptionId});
+    }
   }
 
   async add(ctx) {
-    const {bookId, sellerId, price = '', status = ''} = ctx.request.body;
-    ctx.body = await quoteModel.insert({bookId, sellerId, price, status});
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {bookId, sellerId, price = '', status = ''} = ctx.request.body;
+      ctx.body = await quoteModel.insert({bookId, sellerId, price, status});
+    }
   }
 
 
   async delete(ctx) {
-    const {id} = ctx.params;
-    ctx.body = await quoteModel.deleteById({id})
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {id} = ctx.params;
+      ctx.body = await quoteModel.deleteById({id})
+    }
   }
 
   async update(ctx) {
-    const {id} = ctx.params;
-    const {bookId, sellerId, price = '', status = ''} = ctx.request.body;
-    ctx.body = await quoteModel.update({id, bookId, sellerId, price, status})
+    const {_auth: roleId} = JWTDecode(ctx.header.authorization);
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {id} = ctx.params;
+      const {bookId, sellerId, price = '', status = ''} = ctx.request.body;
+      ctx.body = await quoteModel.update({id, bookId, sellerId, price, status})
+    }
   }
 
   async addQuote(ctx) {
     const {_auth: roleId, _uid: id} = JWTDecode(ctx.header.authorization);
-    const {id: bookId, price = ''} = ctx.request.body;
-
+    const {id: bookId, price = '',subscriptionId} = ctx.request.body;
     if (roleId !== 5) {
       ctx.body = {code: 9999, message: '你无权进行此操作'}
     } else {
-      ctx.body = await quoteModel.insert({sellerId: id, bookId, price, status: 1})
+      ctx.body = await quoteModel.insert({sellerId: id, bookId, price, status: 1,subscriptionId})
+    }
+  }
+
+  async updateQuote(ctx) {
+    const {_auth: roleId, _uid: id} = JWTDecode(ctx.header.authorization);
+    if (roleId !== 5) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {id: bookId} = ctx.params;
+      const {price = '',subscriptionId} = ctx.request.body;
+      ctx.body = await quoteModel.updateQuote({sellerId: id, bookId, price,subscriptionId})
     }
   }
 
@@ -65,8 +96,12 @@ class QuoteController {
 
   async sub(ctx) {
     const {_auth: roleId} = JWTDecode(ctx.header.authorization);
-    const {bookId, sellerId} = ctx.request.body;
-    ctx.body = await quoteModel.sub({sellerId, bookId});
+    if (![1, 2].includes(roleId)) {
+      ctx.body = {code: 9999, message: '你无权进行此操作'}
+    } else {
+      const {bookId, sellerId,subscriptionId} = ctx.request.body;
+      ctx.body = await quoteModel.sub({sellerId, bookId,subscriptionId});
+    }
   }
 }
